style: 고객센터 화면 전환 스타일 수정 #153
- 전환 시 헤더 제외 내용만 슬라이드 되도록 함
