+++ src/features/common/routes/HelpDesk/Select.tsx
@@ -1,8 +1,4 @@
-import { NavArrowLeft } from "iconoir-react";
-import { useNavigate } from "react-router-dom";
-
 import Button from "@/components/Button";
-import Header from "@/components/Layout/Header";
 
 import { Buttons, Container, Content } from "./Select.style";
 
@@ -12,16 +8,8 @@ interface Props {
 }
 
 export default function Select({ inquiryTypeName, onClick }: Props) {
-  const navigate = useNavigate();
-
   return (
     <Container>
-      <Header>
-        <Header.Left>
-          <NavArrowLeft onClick={() => navigate(-1)} />
-        </Header.Left>
-        <Header.Center>고객센터</Header.Center>
-      </Header>
       <Content>
         <div>
           <h1>어떤 문제를 해결해 드릴까요?</h1>
