+++ src/features/common/routes/HelpDesk/index.tsx
@@ -1,7 +1,10 @@
 import styled from "@emotion/styled";
+import { NavArrowLeft } from "iconoir-react";
 import { useState } from "react";
+import { useNavigate } from "react-router";
 
 import Head from "@/components/Head";
+import Header from "@/components/Layout/Header";
 
 import Form from "./Form";
 import Select from "./Select";
@@ -10,18 +13,32 @@ import Success from "./Success";
 export default function HelpDesk() {
   const [inquiryType, setInquiryType] = useState(0);
   const inquiryTypeName = ["기능 추가 건의", "버그 신고", "기타 문의"];
+  const [headerTitle, setHeaderTitle] = useState("고객센터");
   const [translateX, setTranslateX] = useState(0);
 
   const [success, setSuccess] = useState(false);
+  const navigate = useNavigate();
+
+  const CHANGE_HEADER_TITLE_DELAY = 200;
 
   const goPrev = () => {
     setInquiryType(0);
     setTranslateX(0);
+    setTimeout(() => setHeaderTitle("고객센터"), CHANGE_HEADER_TITLE_DELAY);
   };
 
   const handleItemClick = (i: number) => {
     setInquiryType(i + 1);
     setTranslateX(-50);
+    setTimeout(
+      () => setHeaderTitle(inquiryTypeName[i]),
+      CHANGE_HEADER_TITLE_DELAY,
+    );
+  };
+
+  const handlerPrevClick = () => {
+    if (inquiryType !== 0) goPrev();
+    else navigate(-1);
   };
 
   return (
@@ -29,9 +46,16 @@ export default function HelpDesk() {
       <Head title="오덕 | 고객센터" />
       <Container>
         {success && <Success />}
+        <Header>
+          <Header.Left>
+            <NavArrowLeft onClick={handlerPrevClick} />
+          </Header.Left>
+          <Header.Center>{headerTitle}</Header.Center>
+        </Header>
         <Slide translateX={translateX}>
           <Select inquiryTypeName={inquiryTypeName} onClick={handleItemClick} />
           <Form
+            key={inquiryType}
             inquiryTypeName={inquiryTypeName[inquiryType - 1]}
             goPrev={goPrev}
             setSuccess={setSuccess}
@@ -51,7 +75,7 @@ const Container = styled.div`
 
 const Slide = styled.div<{ translateX?: number }>`
   width: 200%;
-  height: 100%;
+  height: calc(100% - 60px);
   display: flex;
   transform: ${({ translateX = 0 }) => `translateX(${translateX}%)`};
   transition: all 0.3s;
